requirements.txt
fastapi
uvicorn[standard]
pandas
prophet
python-dotenv
apscheduler
requests
statsmodels
# tbats  # removed due to build issues
scikit-learn
xgboost
lightgbm
schedule